Remove error in ComBat function call
When trying to duplicate your processing of the curatedOvarianData, I found that the option numCovs=NULL in the call to the ComBat function was throwing the following error:
Error in ComBat(dat = GEN_Data, batch = BatchDataSelected[, batchColName],  : 
  unused argument (numCovs = NULL)
Removing this argument fixes the problem. This may be a versioning issue, but I'll leave it up to you to decide.
Thanks!